fix new tenant migrations

+++ database/migrations/tenant/2026_07_26_100001_add_batch_to_whole_sale_prices_table.php
@@ -7,16 +7,20 @@
     return new class extends Migration {
         public function up() : void
         {
-            Schema::table( 'whole_sale_prices' , function (Blueprint $table) {
-                $table->string( 'batch' )->nullable();
-            } );
-            DB::table( 'whole_sale_prices' )->update( [ 'batch' => time() ] );
+            if ( ! Schema::hasColumn( 'whole_sale_prices' , 'batch' ) ) {
+                Schema::table( 'whole_sale_prices' , function (Blueprint $table) {
+                    $table->string( 'batch' )->nullable();
+                } );
+                DB::table( 'whole_sale_prices' )->update( [ 'batch' => time() ] );
+            }
         }
 
         public function down() : void
         {
-            Schema::table( 'whole_sale_prices' , function (Blueprint $table) {
-                $table->dropColumn( 'batch' );
-            } );
+            if ( Schema::hasColumn( 'whole_sale_prices' , 'batch' ) ) {
+                Schema::table( 'whole_sale_prices' , function (Blueprint $table) {
+                    $table->dropColumn( 'batch' );
+                } );
+            }
         }
     };
